# mempool: Optimize orphan map limiting