--- a/doc.go
+++ b/doc.go
@@ -121,7 +121,7 @@ Application Options:
       --sigcachemaxsize=    The maximum number of entries in the signature
                             verification cache.
       --blocksonly          Do not accept transactions from remote peers.
-      --relaynonstd         Relay non-standard transactions regardless of the
+      --acceptnonstd        Accept non-standard transactions regardless of the
                             default settings for the active network.
       --rejectnonstd        Reject non-standard transactions regardless of the
                             default settings for the active network.